+++ src/Identity/IdentityInterface.php
@@ -18,4 +18,21 @@ interface IdentityInterface
      * @return bool
      */
     public function isAuthenticated(): bool;
+
+    /**
+     * Get attributes.
+     *
+     * @return array<string, mixed>
+     */
+    public function getAttributes(): array;
+
+    /**
+     * Get attribute.
+     *
+     * @param string     $name
+     * @param mixed|null $default
+     *
+     * @return mixed
+     */
+    public function getAttribute(string $name, mixed $default = null): mixed;
 }