+++ tests/Identity/IdentityTest.php
@@ -15,12 +15,19 @@ class IdentityTest extends TestCase
      * @covers \ExtendsSoftware\ExaPHP\Identity\Identity::__construct()
      * @covers \ExtendsSoftware\ExaPHP\Identity\Identity::getIdentifier()
      * @covers \ExtendsSoftware\ExaPHP\Identity\Identity::isAuthenticated()
+     * @covers \ExtendsSoftware\ExaPHP\Identity\Identity::getAttributes()
+     * @covers \ExtendsSoftware\ExaPHP\Identity\Identity::getAttribute()
      */
     public function testGetIdentifier(): void
     {
-        $identity = new Identity('foo', true);
+        $identity = new Identity('foo', true, [
+            'foo' => 'bar',
+        ]);
 
         $this->assertSame('foo', $identity->getIdentifier());
         $this->assertSame(true, $identity->isAuthenticated());
+        $this->assertSame(['foo' => 'bar'], $identity->getAttributes());
+        $this->assertSame('bar', $identity->getAttribute('foo'));
+        $this->assertSame('foo', $identity->getAttribute('bar', 'foo'));
     }
 }